Added success screens to exercise assignment and patient creation
